Move some common test helpers to flocker.testtools
